Updated the lorensen method: remove empty cubes beforehand and correct isolevel handling.

+++ include/lib/lorensen.cuh
@@ -5,7 +5,7 @@
 
 namespace mc {
     namespace lorensen {
-        std::tuple<float *, size_t, int *, size_t> marching_cubes(
+        std::tuple<float *, uint32_t, int *, uint32_t> marching_cubes(
             float * const grid_ptr,  
             const std::array<int64_t, 3>& grid_shape, 
             const std::array<float, 6>& aabb,
